TIMAAT Weiterentwicklung #2 - Implemented transcription editing of music into client
* The music management view now has a separate tab to edit transcriptions
* Inside this tab you can perform crud operations on transcriptions
* Fixed id datatypes and query used to persist music translation information inside the database

+++ src/main/java/de/bitgilde/TIMAAT/model/FIPOP/MusicTranslationPK.java
@@ -18,31 +18,31 @@ public class MusicTranslationPK implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Column(name = "music_id", updatable = false, nullable = false, insertable = false)
-    private Long musicId;
+    private Integer musicId;
     @Column(name = "language_id", updatable = false, nullable = false, insertable = false)
-    private Long languageId;
+    private Integer languageId;
 
     public MusicTranslationPK() {
     }
 
-    public MusicTranslationPK(Long musicId, Long languageId) {
+    public MusicTranslationPK(Integer musicId, Integer languageId) {
         this.musicId = musicId;
         this.languageId = languageId;
     }
 
-    public Long getLanguageId() {
+    public Integer getLanguageId() {
         return languageId;
     }
 
-    public void setLanguageId(Long languageId) {
+    public void setLanguageId(Integer languageId) {
         this.languageId = languageId;
     }
 
-    public Long getMusicId() {
+    public Integer getMusicId() {
         return musicId;
     }
 
-    public void setMusicId(Long musicId) {
+    public void setMusicId(Integer musicId) {
         this.musicId = musicId;
     }
